refactor: extract some functions into posts module
