+++ src/scripts/posts.ts
@@ -1,0 +1,13 @@
+import { MarkdownInstance } from "astro";
+
+export interface PostFrontmatter {
+    title: string;
+    description?: string;
+    tags?: string[];
+};
+
+export type Post = MarkdownInstance<PostFrontmatter>;
+
+export const allPosts: Post[] = Object.values(
+    import.meta.glob("/src/pages/posts/*.{mdx,astro}", { eager: true })
+).map((p) => p as Post);
